rpi/backend/requirements.txt
Flask==0.10.1
Flask-Cors==2.1.0
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
distribute==0.6.24
gunicorn==19.3.0
itsdangerous==0.24
pyserial==2.7
six==1.9.0
waitress==0.8.9
wsgiref==0.1.2
